When extending the compiled transitions using the POSIX-style future rule, search for the last two distinct transition types rather than assuming that the last two transitions already have distinct types.
The 2016g zic started adding a redundant final transition when
the future rule disagreed with the local time type after the
last explicit transition in the data (to avoid a quirk in the
glibc zoneinfo implementation).  This only affects the Morocco
and Troll rules at the moment.

Move the extension logic into its own function for clarity.